% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assemble.R
\name{assemble}
\alias{assemble}
\title{Operation of networks}
\usage{
assemble(..., mode = c("union", "intersection", "difference", "complementer"))
}
\arguments{
\item{...}{A list of igraph objects.}

\item{mode}{Operation mode, possible values are \code{union}, \code{intersection},
\code{difference}, and \code{complementer}.}
}
\value{
An integrated igraph object.
}
\description{
Operation of networks including assemble them into a whole one,
find their intersection, difference or complementer.
}
\examples{
g1<-graph.ring(10)
g2<-graph.tree(10,mode="undirected")
assemble(g1,g2,mode="union")
assemble(g1,g2,mode="intersection")
assemble(g1,g2,mode="difference")
assemble(g1,mode="complementer")
}

