% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ath.R
\docType{data}
\name{ath}
\alias{ath}
\title{Arabidopsis thaliana Biogrid PPI database}
\format{A data frame with 39901 rows and 4 columns.}
\source{
\url{http://thebiogrid.org/}
}
\usage{
data(ath)
}
\description{
The Arabidopsis thaliana protein-protein interactions obtained from the Biogrid database.
}
\references{
Stark C, Breitkreutz BJ, Reguly T, Boucher L, Breitkreutz A, Tyers M (Jan 2006). BioGRID: A General Repository for Interaction Datasets. Nucleic Acids Research 1 (34): 535-539.
}
\keyword{datasets}

