% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/construction.R
\name{construct_local}
\alias{construct_local}
\title{Local network construction}
\usage{
construct_local(input, node.attribute)
}
\arguments{
\item{input}{A data frame containing the experimental data.}

\item{node.attribute}{A data frame containing node attributes.}
}
\value{
A network in igraph format.
}
\description{
Construction a network from experimental data.
}
\examples{
local<-data.frame(1:5,2:6)
attribute<-data.frame(c(0.2,0.3,0.2,0.5,0.1))
net<-construct_local(input=local,node.attribute=attribute)
}
\seealso{
\code{\link{construction}}, \code{\link{construct_nlocal}}.
}

