% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GO_function.R
\docType{data}
\name{GO_function}
\alias{GO_function}
\title{GO function annotation database}
\format{A data frame with 100475 rows and 3 columns.}
\source{
\url{http://geneontology.org/}
}
\usage{
data(GO_function)
}
\description{
The gene products' GO function annoation data retrieved from the Gene Ontology Consortium.
}
\references{
The Gene Ontology Consortium (January 2008). The Gene Ontology project in 2008. Nucleic Acids Res. 36 (Database issue): 440-4.
}
\keyword{datasets}

