% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GO_process.R
\docType{data}
\name{GO_process}
\alias{GO_process}
\title{GO process annotation database}
\format{A data frame with 128103 rows and 3 columns.}
\source{
\url{http://geneontology.org/}
}
\usage{
data(GO_process)
}
\description{
The gene products' GO process annoation data retrieved from the Gene Ontology Consortium.
}
\references{
The Gene Ontology Consortium (January 2008). The Gene Ontology project in 2008. Nucleic Acids Res. 36 (Database issue): 440-4.
}
\keyword{datasets}

