% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assemble.R
\name{graph.add}
\alias{graph.add}
\title{Network union}
\usage{
graph.add(...)
}
\arguments{
\item{...}{A list of igraph objects.}
}
\value{
An integrated igraph object.
}
\description{
Assemble a list of networks into a whole one.
}
\examples{
g1<-graph.ring(10)
g2<-graph.tree(10,mode="undirected")
g<-graph.add(g1,g2)
}

