% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HPRD.R
\docType{data}
\name{HPRD}
\alias{HPRD}
\title{HPRD PPI database}
\format{A data frame with 39240 rows and 2 columns.}
\source{
\url{http://www.hprd.org/}
}
\usage{
data(HPRD)
}
\description{
The human protein-protein interactions obtained from the HPRD database.
}
\references{
Peri S, et al. (2003). Development of human protein reference database as an initial platform for approaching systems biology in humans. Genome Research 13: 2363-71.
}
\keyword{datasets}

