% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/human.R
\docType{data}
\name{human}
\alias{human}
\title{Homo sapiens Biogrid PPI database}
\format{A data frame with 250802 rows and 4 columns.}
\source{
\url{http://thebiogrid.org/}
}
\usage{
data(human)
}
\description{
The Homo sapien protein-protein interactions obtained from the Biogrid database.
}
\references{
Stark C, Breitkreutz BJ, Reguly T, Boucher L, Breitkreutz A, Tyers M (Jan 2006). BioGRID: A General Repository for Interaction Datasets. Nucleic Acids Research 1 (34): 535-539.
}
\keyword{datasets}

