% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locDB.R
\docType{data}
\name{locDB}
\alias{locDB}
\title{Gene products' location database}
\format{A list of 2 data frames.}
\source{
\url{https://www.rostlab.org/services/locDB/}
}
\usage{
data(locDB)
}
\description{
The human's and ath's gene products' subcellular location data.
}
\references{
Rastogi, Shruti; Rost Burkhard (Jan 2011). LocDB: experimental annotations of localization for Homo sapiens and Arabidopsis thaliana. Nucleic Acids Res. 39 (Database issue): 230-4.
}
\keyword{datasets}

