\name{ProNet-package}
\alias{ProNet-package}
\alias{ProNet}
\docType{package}
\title{
Biological Network Construction, Visulization and Analyses
}
\description{
High-throughput experiments are now widely used in biological researches, which improves both the quality and quantity of omics data. Network-based presentation of these data has become a popular way in data analyses. This package mainly provides functions for biological network construction, visulization and analyses. Networks can be constructed either from experimental data or from a set of proteins and integrated PPI database. Based on them, users can perform traditional visulization, along with the subcellular location based ones for Homo spaiens and Arabidopsis thaliana. Furthermore, analyses including topological statistics, functional module clustering and go profiling can also be achieved. 
}
\details{
\tabular{ll}{
Package: \tab ProNet\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2015-10-16\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Xiang-Yun Wu and Xia-Yu Xia

Maintainer: Xia-Yu Xia <xiaxiayu.thu@hotmail.com>
}
\keyword{ package }


