% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topology.R
\name{topology_anc}
\alias{topology_anc}
\title{Average neighborhood connectivity statistics}
\usage{
topology_anc(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\value{
A data frame containing the average neighborhood connectivity information and plots.
}
\description{
Average neighborhood connectivity statistics of the network.
}
\examples{
nlocal<-data.frame(c("DVL1","DVL2","DVL3"))
net<-construction(input=nlocal,db="HPRD",species="human",ID.type="Gene symbol",hierarchy=1)
anc<-topology_anc(net)
}

