% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topology.R
\name{topology_betweenness}
\alias{topology_betweenness}
\title{Betweenness statistics}
\usage{
topology_betweenness(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\value{
A data frame containing the betweenness distribution and plots.
}
\description{
Betweenness statistics of the network.
}
\examples{
nlocal<-data.frame(c("DVL1","DVL2","DVL3"))
net<-construction(input=nlocal,db="HPRD",species="human",ID.type="Gene symbol",hierarchy=1)
b.g<-topology_betweenness(net)
}

