% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topology.R
\name{topology_closeness}
\alias{topology_closeness}
\title{Closeness statistics}
\usage{
topology_closeness(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\value{
A data frame containing the closeness distribution and plots.
}
\description{
Closeness statistics of the network.
}
\examples{
nlocal<-data.frame(c("DVL1","DVL2","DVL3"))
net<-construction(input=nlocal,db="HPRD",species="human",ID.type="Gene symbol",hierarchy=1)
c<-topology_closeness(net)
}

