% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topology.R
\name{topology_cluster_coeffi}
\alias{topology_cluster_coeffi}
\title{Clustering coefficient statistics}
\usage{
topology_cluster_coeffi(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\value{
A data frame containing the clustering coefficient information and plots.
}
\description{
Clustering coefficient statistics of the network.
}
\examples{
nlocal<-data.frame(c("DVL1","DVL2","DVL3"))
net<-construction(input=nlocal,db="HPRD",species="human",ID.type="Gene symbol",hierarchy=1)
cc<-topology_cluster_coeffi(net)
}

