% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topology.R
\name{topology_ec}
\alias{topology_ec}
\title{Eigenvector centrality statistics}
\usage{
topology_ec(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\value{
A data frame containing the vertex eigenvector centrality information and plots.
}
\description{
Eigenvector centrality statistics of the network.
}
\examples{
nlocal<-data.frame(c("DVL1","DVL2","DVL3"))
net<-construction(input=nlocal,db="HPRD",species="human",ID.type="Gene symbol",hierarchy=1)
ec<-topology_ec(net)
}

