% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topology.R
\name{topology_shortest_paths}
\alias{topology_shortest_paths}
\title{Shortest path statistics}
\usage{
topology_shortest_paths(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\value{
An array containing the vertex path information and plots.
}
\description{
Shortest path statistics of the network.
}
\examples{
nlocal<-data.frame(c("DVL1","DVL2","DVL3"))
net<-construction(input=nlocal,db="HPRD",species="human",ID.type="Gene symbol",hierarchy=1)
p<-topology_shortest_paths(net)
}

