% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topology.R
\name{topology_simple}
\alias{topology_simple}
\title{Basic statistics}
\usage{
topology_simple(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\value{
A vector containing the simple statistics.
}
\description{
Basic statistics of the network's topological parameters.
}
\examples{
nlocal<-data.frame(c("DVL1","DVL2","DVL3"))
net<-construction(input=nlocal,db="HPRD",species="human",ID.type="Gene symbol",hierarchy=1)
s<-topology_simple(net)
}

